Regression Model on Market Segmentation
Course: Advanced Statistics

This project identified the factors to study satisfaction level of customer. A regression model to be built using these factors for Market segmentation.The objective of the project is to use the dataset to build a regression model to predict customer satisfaction.

Project Objective
The project is based on Multi-Variate analysis and its application in Market Segmentation in context of Product Service management. The objective is to check the given data-set and variables for multicollinearity, perform factor analysis and then build a multiple regression model to predict Customer Satisfaction based of the four factors chosen during factor analysis.

What is multicollinearity and how it affects the regression model? Multicollinearity occurs when the independent variables of a regression model are correlated and if the degree of collinearity between the independent variables are high, it becomes difficult to estimate the relationship between each independent variable and the dependent variable and the overall precision of the estimate coefficients. Even though the regression models with high multicollinearity can give you a high R squared but hardly any significant variables

Assumptions
1) There should not be multicollinearity between the variables.
2) The variance of residual is same is that of dependent variable is Satisfaction 3) Linearity exists between dependent and independent variables
4) Observations are independent to each other.